Create PUNT.html
Basically, a TO-DO (Later) List.  Punt a block down a football field that is the visual metaphor for "I'll do it later".  

Each 'football field' can be scaled to either a week or full month or even a day (toggles at the top; repopulates data and scaling)

UI: topbar:searchfield; bottombar:Just a chatText entry to start typing note (enter or button-click); produces a note (consider shorthand in text entry to dicate data parameters (for instance, like a subreddit /r: "h/16" sets power and trajectory to launch block 16 hours (relative to the month scale); so when you tap the block to launch it (essentially interactive psuedo-automation), it goes where intended, without having to tap additional UI; otherwise, place it manually as so desired.


shorthand for h/24h could be "hard stop in 24 hours", imposing virtual impositions on user placement/launch in placing the cube.